fix quantity

+++ src/components/Products/products.jsx
@@ -91,12 +91,20 @@ const Products = ({title,products,filterBySubcats,filterByPrice,selectPrice,setS
                     <div className="products__item__miniature">
                         <span>{product.origin}</span>
                         <div className={`products__item__add__btn ${activeBtn.some(item => item === i) ? "active" : ""}`} onClick={() => {
-                            setActiveBtn([...activeBtn, i])       
+                            if(!product.quantity === 0){
+                                setActiveBtn([...activeBtn, i]) 
+                                return
+                            }
+                            setActiveBtn([...activeBtn])
                         }}>
-                            {activeBtn.some(item => item === i) && <span className='icon__btn__add' onClick={() =>{
-                                product.quantity--
-                                dispatch(minusQuantity(product))
-                                dispatch(deleteItemBasket(product))
+                            {activeBtn.some(item => item === i) && <span className='icon__btn__add' onClick={(e) =>{
+                                if(product.quantity !== 0){
+                                    product.quantity--
+                                    dispatch(minusQuantity(product))
+                                    dispatch(deleteItemBasket(product))
+                                    return
+                                }
+                                
                             }}><Icon path={mdiMinus} size={1} color={'white'} /></span>}
 
                             {activeBtn.some(item => item === i) && <div className='product__quantity'>{product.quantity}</div>}
