+++ src/pages/Main/mainPage.jsx
@@ -273,9 +273,12 @@ const MainPage = () =>{
                             setActiveBtn([...activeBtn, i])       
                         }}>
                             {activeBtn.some(item => item === i) && <span className='icon__btn__add' onClick={() =>{
-                                product.quantity--
-                                dispatch(minusQuantity(product))
-                                dispatch(deleteItemBasket(product))
+                                if(product.quantity !== 0){
+                                    product.quantity--
+                                    dispatch(minusQuantity(product))
+                                    dispatch(deleteItemBasket(product))
+                                    return
+                                }
                             }}><Icon path={mdiMinus} size={1} color={'white'} /></span>}
 
                             {activeBtn.some(item => item === i) && <div className='product__quantity'>{product.quantity}</div>}
@@ -405,9 +408,12 @@ const MainPage = () =>{
                             setActiveBtn([...activeBtn, i])       
                         }}>
                             {activeBtn.some(item => item === i) && <span className='icon__btn__add' onClick={() =>{
+                               if(product.quantity !== 0){
                                 product.quantity--
                                 dispatch(minusQuantity(product))
                                 dispatch(deleteItemBasket(product))
+                                return
+                            }
                             }}><Icon path={mdiMinus} size={1} color={'white'} /></span>}
 
                             {activeBtn.some(item => item === i) && <div className='product__quantity'>{product.quantity}</div>}
@@ -538,9 +544,12 @@ const MainPage = () =>{
                             setActiveBtn([...activeBtn, i])       
                         }}>
                             {activeBtn.some(item => item === i) && <span className='icon__btn__add' onClick={() =>{
-                                product.quantity--
-                                dispatch(minusQuantity(product))
-                                dispatch(deleteItemBasket(product))
+                                if(product.quantity !== 0){
+                                    product.quantity--
+                                    dispatch(minusQuantity(product))
+                                    dispatch(deleteItemBasket(product))
+                                    return
+                                }
                             }}><Icon path={mdiMinus} size={1} color={'white'} /></span>}
 
                             {activeBtn.some(item => item === i) && <div className='product__quantity'>{product.quantity}</div>}
@@ -670,9 +679,12 @@ const MainPage = () =>{
                             setActiveBtn([...activeBtn, i])       
                         }}>
                             {activeBtn.some(item => item === i) && <span className='icon__btn__add' onClick={() =>{
-                                product.quantity--
-                                dispatch(minusQuantity(product))
-                                dispatch(deleteItemBasket(product))
+                                if(product.quantity !== 0){
+                                    product.quantity--
+                                    dispatch(minusQuantity(product))
+                                    dispatch(deleteItemBasket(product))
+                                    return
+                                }
                             }}><Icon path={mdiMinus} size={1} color={'white'} /></span>}
 
                             {activeBtn.some(item => item === i) && <div className='product__quantity'>{product.quantity}</div>}
@@ -782,9 +794,12 @@ const MainPage = () =>{
                             setActiveBtn([...activeBtn, i])       
                         }}>
                             {activeBtn.some(item => item === i) && <span className='icon__btn__add' onClick={() =>{
+                               if(product.quantity !== 0){
                                 product.quantity--
                                 dispatch(minusQuantity(product))
                                 dispatch(deleteItemBasket(product))
+                                return
+                            }
                             }}><Icon path={mdiMinus} size={1} color={'white'} /></span>}
 
                             {activeBtn.some(item => item === i) && <div className='product__quantity'>{product.quantity}</div>}
@@ -893,9 +908,12 @@ const MainPage = () =>{
                             setActiveBtn([...activeBtn, i])       
                         }}>
                             {activeBtn.some(item => item === i) && <span className='icon__btn__add' onClick={() =>{
-                                product.quantity--
-                                dispatch(minusQuantity(product))
-                                dispatch(deleteItemBasket(product))
+                                if(product.quantity !== 0){
+                                    product.quantity--
+                                    dispatch(minusQuantity(product))
+                                    dispatch(deleteItemBasket(product))
+                                    return
+                                }
                             }}><Icon path={mdiMinus} size={1} color={'white'} /></span>}
 
                             {activeBtn.some(item => item === i) && <div className='product__quantity'>{product.quantity}</div>}
@@ -1026,9 +1044,12 @@ const MainPage = () =>{
                             setActiveBtn([...activeBtn, i])       
                         }}>
                             {activeBtn.some(item => item === i) && <span className='icon__btn__add' onClick={() =>{
-                                product.quantity--
-                                dispatch(minusQuantity(product))
-                                dispatch(deleteItemBasket(product))
+                                if(product.quantity !== 0){
+                                    product.quantity--
+                                    dispatch(minusQuantity(product))
+                                    dispatch(deleteItemBasket(product))
+                                    return
+                                }
                             }}><Icon path={mdiMinus} size={1} color={'white'} /></span>}
 
                             {activeBtn.some(item => item === i) && <div className='product__quantity'>{product.quantity}</div>}
@@ -1156,9 +1177,12 @@ const MainPage = () =>{
                             setActiveBtn([...activeBtn, i])       
                         }}>
                             {activeBtn.some(item => item === i) && <span className='icon__btn__add' onClick={() =>{
+                               if(product.quantity !== 0){
                                 product.quantity--
                                 dispatch(minusQuantity(product))
                                 dispatch(deleteItemBasket(product))
+                                return
+                            }
                             }}><Icon path={mdiMinus} size={1} color={'white'} /></span>}
 
                             {activeBtn.some(item => item === i) && <div className='product__quantity'>{product.quantity}</div>}
@@ -1287,9 +1311,12 @@ const MainPage = () =>{
                             setActiveBtn([...activeBtn, i])       
                         }}>
                             {activeBtn.some(item => item === i) && <span className='icon__btn__add' onClick={() =>{
+                               if(product.quantity !== 0){
                                 product.quantity--
                                 dispatch(minusQuantity(product))
                                 dispatch(deleteItemBasket(product))
+                                return
+                            }
                             }}><Icon path={mdiMinus} size={1} color={'white'} /></span>}
 
                             {activeBtn.some(item => item === i) && <div className='product__quantity'>{product.quantity}</div>}
@@ -1417,9 +1444,12 @@ const MainPage = () =>{
                             setActiveBtn([...activeBtn, i])       
                         }}>
                             {activeBtn.some(item => item === i) && <span className='icon__btn__add' onClick={() =>{
+                               if(product.quantity !== 0){
                                 product.quantity--
                                 dispatch(minusQuantity(product))
                                 dispatch(deleteItemBasket(product))
+                                return
+                            }
                             }}><Icon path={mdiMinus} size={1} color={'white'} /></span>}
 
                             {activeBtn.some(item => item === i) && <div className='product__quantity'>{product.quantity}</div>}
@@ -1547,9 +1577,12 @@ const MainPage = () =>{
                             setActiveBtn([...activeBtn, i])       
                         }}>
                             {activeBtn.some(item => item === i) && <span className='icon__btn__add' onClick={() =>{
-                                product.quantity--
-                                dispatch(minusQuantity(product))
-                                dispatch(deleteItemBasket(product))
+                                if(product.quantity !== 0){
+                                    product.quantity--
+                                    dispatch(minusQuantity(product))
+                                    dispatch(deleteItemBasket(product))
+                                    return
+                                }
                             }}><Icon path={mdiMinus} size={1} color={'white'} /></span>}
 
                             {activeBtn.some(item => item === i) && <div className='product__quantity'>{product.quantity}</div>}
